#255 [QA] 헬스케어 오늘의 걸음수 업로드 시 소수점 절삭 되어 보내짐

+++ core/data/src/main/java/com/startup/data/remote/datasourceimpl/StepDataStoreImpl.kt
@@ -196,7 +196,7 @@ class StepDataStoreImpl @Inject constructor(
     }
 
     private suspend fun putHealthcareInfo(targetDate: Long, currentSteps: Int) {
-        val distanceMeters = (round((0.0006 * currentSteps) * 100) / 100.0).toInt()
+        val distanceMeters = (round((0.0006 * currentSteps) * 100) / 100.0)
         val calories = currentSteps / 30
         if (DateUtil.isSameDay(targetDate, System.currentTimeMillis())) {
             return
